requirements.txt
archspec==0.2.5
asttokens==3.0.0
astunparse==1.6.3
attrs==25.1.0
beautifulsoup4==4.12.3
boltons==24.0.0
bottle==0.12.25
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.4
colorama==0.4.6
coloredlogs==15.0.1
conan==1.66.0
conda==25.1.0
conda-build==25.1.1
conda_index==0.5.0
conda-libmamba-solver==25.1.1
conda-package-handling==2.4.0
conda_package_streaming==0.11.0
decorator==5.1.1
distro==1.8.0
dnspython==2.7.0
docutils==0.21.2
exceptiongroup==1.2.2
executing==2.1.0
expecttest==0.3.0
fasteners==0.20
filelock==3.17.0
flatbuffers==25.2.10
frozendict==2.4.6
fsspec==2024.12.0
h2==4.1.0
hpack==4.0.0
humanfriendly==10.0
hyperframe==6.0.1
hypothesis==6.124.7
idna==3.10
importlib_resources==6.5.2
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
libarchive-c==5.1
libmambapy==2.0.5
lief==0.14.1
lintrunner==0.12.7
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
menuinst==2.2.0
more-itertools==10.6.0
mpmath==1.3.0
netron==8.5.3
networkx==3.4.2
ninja==1.11.1.3
node-semver==0.6.1
numpy==2.2.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
nvidia-pyindex==1.0.9
onnx==1.18.0
onnx_graphsurgeon==0.5.8
onnxruntime-gpu==1.22.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
optree==0.14.0
packaging==24.2
parso==0.8.4
patch-ng==1.17.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==25.2
pkginfo==1.12.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pluggy==1.5.0
pluginbase==1.0.1
polygraphy==0.49.26
prompt_toolkit==3.0.50
protobuf==6.32.0
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycosat==0.6.6
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-etcd==0.4.5
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.8
setuptools==75.8.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.5
stack_data==0.6.3
sympy==1.13.1
tensorrt_cu12==10.13.2.6
tensorrt_cu12_bindings==10.13.2.6
tensorrt_cu12_libs==10.13.2.6
tensorrt_dispatch_cu12==10.13.2.6
tensorrt_dispatch_cu12_bindings==10.13.2.6
tensorrt_dispatch_cu12_libs==10.13.2.6
tensorrt_lean_cu12==10.13.2.6
tensorrt_lean_cu12_bindings==10.13.2.6
tensorrt_lean_cu12_libs==10.13.2.6
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchelastic==0.2.2
torchvision==0.21.0+cu126
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
truststore==0.10.0
types-dataclasses==0.6.6
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
wheel==0.45.1
zipp==3.21.0
zstandard==0.23.0
